Refactor output module and remove deprecated files
- Updated `output.py` to enhance schedule output generation, including both scheduled and unscheduled classes.
- Removed `output_v2.py` as its functionality has been integrated into `output.py`.
- Updated imports in `room_scheduler.py` and `scheduler.py` to reference the correct data loader and output module.
- Deleted `solver.py` as it is no longer needed in the current architecture.
